fix:Renamed Doctypes
