+++ versa_system/fixtures/workflow.json
@@ -252,7 +252,7 @@
  {
   "docstatus": 0,
   "doctype": "Workflow",
-  "document_type": "Feasibility Property Check",
+  "document_type": "Feasibility Check",
   "is_active": 1,
   "modified": "2024-06-03 10:50:22.214351",
   "name": "Feasibility",
