+++ versa_system/versa_system/doctype/feasibility_check/test_feasibility_check.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestFeasibilityPropertyCheck(FrappeTestCase):
+class TestFeasibilityCheck(FrappeTestCase):
 	pass
