--- versa_system/versa_system/doctype/feasibility_property_check/feasibility_property_check.json
@@ -1,53 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "field:from_lead",
- "creation": "2024-06-03 09:40:11.108634",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "from_lead",
-  "properties"
- ],
- "fields": [
-  {
-   "fieldname": "properties",
-   "fieldtype": "Table",
-   "label": "Properties",
-   "options": "Feasibility Check",
-   "read_only": 1
-  },
-  {
-   "fieldname": "from_lead",
-   "fieldtype": "Link",
-   "label": "From lead",
-   "options": "Lead",
-   "unique": 1
-  }
- ],
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2024-06-07 10:23:38.511565",
- "modified_by": "Administrator",
- "module": "Versa System",
- "name": "Feasibility Property Check",
- "naming_rule": "By fieldname",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
